Add files via upload
Extensive changes to Train. The "best model" is now chosen based on a moving average of a loss ratio and the accuracy of positive samples (pos_acc). Created a seed function for reproducibility. Added new model: an EfficientNetB6 model. There was a typo in the previous Main: I was not using the sampler. Also, positive sample loss is weighted to about 0.4*57. My experiments show that higher values are too aggressive, although I still need to fine tune. We need the 2019 data. I added a patience threshold to train(), although I don't think we'll be using it because after about 2-3 epochs, the model begins overfitting.